+++ CHANGELOG.md
@@ -5,12 +5,15 @@
 ## main
 ### Added
 * default.nix (#513)
+* benchmark (#515)
 ### Fixed
 * Logメッセージをwstringで渡した際のメモリ管理のミスを修正 (#510)
 ### Changed
 * 依存ライブラリの更新 (#509)
 	* crowの更新でコードが変わっています (ver2.9以前が使用していたコミットのcrowではビルドできない)
 * MacOS以外でもcurl8.10〜8.12に対してはビルド前にエラーを出すようにした
+* DataStore2::data_sendをmapからvectorに変更、sync()を1割高速化 (#515)
+* std::condition_variableを使わずpollingするようにし、sync()を3割高速化 (#515)
 
 ## [2.9.0] - 2025-02-09
 ### Added